refactor address management: rename addUpdateAddress to addAddress and replace modal component
